requirements.txt
aws-cdk-lib==2.134.0
constructs==10.3.0
pytest==8.1.1
python-dotenv==1.0.1
boto3==1.34.73
smart-open==7.0.4
dataclasses_json==0.6.4
